Moved stuff around to make things more intuitive.
