+++ lib/constants.dart
@@ -1,3 +1,8 @@
 class Constants {
   static const databaseUpdateKey = "__database_update__";
+  // For Text Resizing Optimization of large Font Sizes
+  static const List<double> stdFontSizes =
+  [1536, 1292, 1086, 914, 768, 646, 456, 322, 272, 228, 192, 162, 136, 114,
+    96, 84, 72, 60, 48, 36, 30, 24, 21, 18, 16, 14, 12, 11, 10, 9, 8, 7, 6];
+
 }
